Exam bug fixes

+++ scenes/1.5 Administration/index.js
@@ -28,7 +28,6 @@ const AdministrationScene = new Scenes.WizardScene("ADMINISTRATION_SCENE",
         }
         const users = (await UserSchema.find({"exams.0": {$exists: true}}).populate('exams').lean()).map(user => {return {...user, exams: [user.exams.sort((a, b) => (a.result < b.result) ? 1: (b.result < a.result) ? -1 : 0)[0]]}})
         ctx.session.state.users = users;
-        console.log(ctx.session.state)
         const result = users.map((user, index) => {
             return `${index + 1}. ${user.job} ${user.name}\nРезультат: ${user.exams[0].result} Дата: ${new Date().toISOString(user.exams[0].result).slice(0, 10).split('-').join('.')}`
         }).join('\n')
@@ -53,10 +52,20 @@ const AdministrationScene = new Scenes.WizardScene("ADMINISTRATION_SCENE",
 
 const GuardScene = new Scenes.WizardScene("GUARD_SCENE", 
     (ctx) => {
-        ctx.reply('Введите ключ.')
+        ctx.reply('Введите ключ.', Markup
+            .keyboard([
+                ['В начало', 'Назад']
+            ])
+            .oneTime()
+            .resize()
+        )
         return ctx.wizard.next();
     },
     async (ctx) => {
+        switch(ctx.message?.text) {
+            case 'Назад': return ctx.scene.enter('MAIN_SCENE'); 
+            case 'В начало': return ctx.scene.enter('MAIN_SCENE'); 
+        }
         if(ctx.message?.text == process.env.HASH) {
             ctx.session.state.op = true
             return ctx.scene.enter('ADMINISTRATION_SCENE')
@@ -72,11 +81,14 @@ const DeleteResultScene = new Scenes.WizardScene("DELETE_RESULT_SCENE",
         return ctx.wizard.next();
     },
     async (ctx) => {
-        if (!isNumber(ctx.message?.text)) ctx.scene.reenter()
+        if (!isNumber(ctx.message?.text)) return ctx.wizard.selectStep(0)
         const number = parseInt(ctx.message.text, 10)
+        if(number >= ctx.session.state.users.length) {
+            console.log('g')
+            return ctx.wizard.selectStep(0)
+        }
         const result = await UserSchema.findByIdAndDelete(ctx.session.state.users[number - 1]._id)
-        console.log(result)        
-        ctx.scene.enter()
+        ctx.scene.enter('ADMINISTRATION_SCENE')
     }
 )
